# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Survivor Guessing Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  player: "player",
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  role: UserProfileRole;
  createdAt: string;
}

export type UpdateRoleBodyRole =
  (typeof UpdateRoleBodyRole)[keyof typeof UpdateRoleBodyRole];

export const UpdateRoleBodyRole = {
  admin: "admin",
  player: "player",
} as const;

export interface UpdateRoleBody {
  role: UpdateRoleBodyRole;
}

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

export const GameStatus = {
  setup: "setup",
  active: "active",
  completed: "completed",
} as const;

export interface Game {
  id: number;
  name: string;
  status: GameStatus;
  /** @nullable */
  survivorWinnerContestantId: number | null;
  currentWeekNumber: number;
  totalWeeks: number;
  createdAt: string;
}

export interface CreateGameBody {
  name: string;
  totalWeeks: number;
}

export type UpdateGameBodyStatus =
  (typeof UpdateGameBodyStatus)[keyof typeof UpdateGameBodyStatus];

export const UpdateGameBodyStatus = {
  setup: "setup",
  active: "active",
  completed: "completed",
} as const;

export interface UpdateGameBody {
  name?: string;
  status?: UpdateGameBodyStatus;
}

export interface Contestant {
  id: number;
  gameId: number;
  name: string;
  createdAt: string;
}

export interface CreateContestantBody {
  name: string;
}

export interface Week {
  id: number;
  gameId: number;
  weekNumber: number;
  isLocked: boolean;
  isOpen: boolean;
  createdAt: string;
}

export interface CreateWeekBody {
  weekNumber: number;
}

export interface Choice {
  id: number;
  questionId: number;
  choiceText: string;
}

export interface QuestionWithChoices {
  id: number;
  weekId: number;
  text: string;
  pointValue: number;
  choices: Choice[];
}

export interface CreateQuestionBody {
  text: string;
  pointValue: number;
  choices?: string[];
}

export interface UpdateQuestionBody {
  text?: string;
  pointValue?: number;
}

export interface CreateChoiceBody {
  choiceText: string;
}

export interface CorrectAnswer {
  id: number;
  questionId: number;
  choiceId: number;
  choiceText: string;
  questionText: string;
  pointValue: number;
}

export interface CorrectAnswerInput {
  questionId: number;
  choiceId: number;
}

export interface SubmitCorrectAnswersBody {
  answers: CorrectAnswerInput[];
}

export interface PlayerAnswer {
  id: number;
  userId: number;
  questionId: number;
  choiceId: number;
  choiceText: string;
  /** @nullable */
  isCorrect: boolean | null;
}

export interface PlayerAnswerInput {
  questionId: number;
  choiceId: number;
}

export interface SaveAnswersBody {
  answers: PlayerAnswerInput[];
}

export interface SurvivorPicks {
  /** @nullable */
  id: number | null;
  /** @nullable */
  userId: number | null;
  gameId: number;
  /** @nullable */
  firstChoiceContestantId: number | null;
  /** @nullable */
  secondChoiceContestantId: number | null;
  /** @nullable */
  firstChoiceName: string | null;
  /** @nullable */
  secondChoiceName: string | null;
  isLocked: boolean;
}

export interface SaveSurvivorPicksBody {
  firstChoiceContestantId: number;
  secondChoiceContestantId: number;
}

export interface SubmitSurvivorWinnerBody {
  winnerContestantId: number;
}

export interface WeeklyPoints {
  weekNumber: number;
  points: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  totalPoints: number;
  weeklyPoints: WeeklyPoints[];
  survivorPickPoints: number;
}

export interface GameStats {
  totalPlayers: number;
  lockedWeeks: number;
  openWeeks: number;
  /** @nullable */
  topPlayer: string | null;
  /** @nullable */
  topPlayerPoints: number | null;
}