# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Survivor Guessing Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games/{gameId}/contestants:
    get:
      operationId: listContestants
      tags: [contestants]
      summary: List contestants for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of contestants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contestant"
    post:
      operationId: createContestant
      tags: [contestants]
      summary: Add a contestant to a game (admin only)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContestantBody"
      responses:
        "201":
          description: Created contestant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contestant"
  /contestants/{contestantId}:
    delete:
      operationId: deleteContestant
      tags: [contestants]
      summary: Delete a contestant (admin only)
      parameters:
        - name: contestantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /weeks/{weekId}/correct-answers:
    get:
      operationId: getCorrectAnswers
      tags: [correct-answers]
      summary: Get correct answers for a week (visible after week is locked)
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Correct answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorrectAnswer"
    post:
      operationId: submitCorrectAnswers
      tags: [correct-answers]
      summary: Submit correct answers for a week (admin only, locks the week)
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCorrectAnswersBody"
      responses:
        "200":
          description: Week scored and locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
  /weeks/{weekId}/my-answers:
    get:
      operationId: getMyAnswers
      tags: [answers]
      summary: Get current player's answers for a week
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerAnswer"
    post:
      operationId: saveMyAnswers
      tags: [answers]
      summary: Save current player's answers for a week
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAnswersBody"
      responses:
        "200":
          description: Saved answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerAnswer"
  /games/{gameId}/survivor-picks:
    get:
      operationId: getMySurvivorPicks
      tags: [survivor-picks]
      summary: Get current player's survivor picks for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survivor picks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurvivorPicks"
    post:
      operationId: saveSurvivorPicks
      tags: [survivor-picks]
      summary: Save current player's survivor picks
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSurvivorPicksBody"
      responses:
        "200":
          description: Saved survivor picks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurvivorPicks"
  /games/{gameId}/survivor-winner:
    post:
      operationId: submitSurvivorWinner
      tags: [survivor-picks]
      summary: Submit the actual Survivor winner (admin only, finalizes game)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitSurvivorWinnerBody"
      responses:
        "200":
          description: Winner submitted and game completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
components:
  schemas:
...
    Game:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [setup, active, completed]
        survivorWinnerContestantId:
          type: ["integer", "null"]
        currentWeekNumber:
          type: integer
        totalWeeks:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - status
        - survivorWinnerContestantId
        - currentWeekNumber
        - totalWeeks
        - createdAt
...
    Contestant:
      type: object
      properties:
        id:
          type: integer
        gameId:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - gameId
        - name
        - createdAt
    CreateContestantBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Week:
      type: object
      properties:
        id:
          type: integer
        gameId:
          type: integer
        weekNumber:
          type: integer
        isLocked:
          type: boolean
        isOpen:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - gameId
        - weekNumber
        - isLocked
        - isOpen
        - createdAt
...
    CorrectAnswer:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        choiceId:
          type: integer
        choiceText:
          type: string
        questionText:
          type: string
        pointValue:
          type: integer
      required:
        - id
        - questionId
        - choiceId
        - choiceText
        - questionText
        - pointValue
    SubmitCorrectAnswersBody:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/CorrectAnswerInput"
      required:
        - answers
    CorrectAnswerInput:
      type: object
      properties:
        questionId:
          type: integer
        choiceId:
          type: integer
      required:
        - questionId
        - choiceId
    PlayerAnswer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        questionId:
          type: integer
        choiceId:
          type: integer
        choiceText:
          type: string
        isCorrect:
          type: ["boolean", "null"]
      required:
        - id
        - userId
        - questionId
        - choiceId
        - choiceText
        - isCorrect
    SaveAnswersBody:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/PlayerAnswerInput"
      required:
        - answers
    PlayerAnswerInput:
      type: object
      properties:
        questionId:
          type: integer
        choiceId:
          type: integer
      required:
        - questionId
        - choiceId
    SurvivorPicks:
      type: object
      properties:
        id:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        gameId:
          type: integer
        firstChoiceContestantId:
          type: ["integer", "null"]
        secondChoiceContestantId:
          type: ["integer", "null"]
        firstChoiceName:
          type: ["string", "null"]
        secondChoiceName:
          type: ["string", "null"]
        isLocked:
          type: boolean
      required:
        - id
        - userId
        - gameId
        - firstChoiceContestantId
        - secondChoiceContestantId
        - firstChoiceName
        - secondChoiceName
        - isLocked
    SaveSurvivorPicksBody:
      type: object
      properties:
        firstChoiceContestantId:
          type: integer
        secondChoiceContestantId:
          type: integer
      required:
        - firstChoiceContestantId
        - secondChoiceContestantId
    SubmitSurvivorWinnerBody:
      type: object
      properties:
        winnerContestantId:
          type: integer
      required:
        - winnerContestantId
...