# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Survivor Guessing Game API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from "./gameStatus";

export interface Game {
  id: number;
  name: string;
  status: GameStatus;
  /** @nullable */
  survivorWinnerContestantId: number | null;
  currentWeekNumber: number;
  totalWeeks: number;
  createdAt: string;
}