# lib/api-zod/src/generated/types/survivorPicks.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Survivor Guessing Game API
 * OpenAPI spec version: 0.1.0
 */

export interface SurvivorPicks {
  /** @nullable */
  id: number | null;
  /** @nullable */
  userId: number | null;
  gameId: number;
  /** @nullable */
  firstChoiceContestantId: number | null;
  /** @nullable */
  secondChoiceContestantId: number | null;
  /** @nullable */
  firstChoiceName: string | null;
  /** @nullable */
  secondChoiceName: string | null;
  isLocked: boolean;
}